Add neovim plugins
After profiling neovim startup quite a lot I've come to the conclusion
that there are more or less two modes of operation. You can run almost
vanilla and have a startup time of less than 100ms or you can add a ton
of plugins and have a startup time of 200-500ms. I'll take the slower
startup for the added convenience of plugins.

I can write code with vanilla vim but I choose not to