Fix man on MacOS
programs.man.enable = true is not necessary and there are some issues
whhere it's advised to set this to false. But the thing that really
matters is having a proper manpath. Can't use ~ for example.

I'm still puzzled why I need to mess around with the MANPATH env var at
all.