Fix VSCode extension issue
This should hopefully fix the issue once and for all. The problem
appears to be that I was downloading extensions using the "latest" URL
but then I extracted the version and stored it in the .nix file. As a
result, when I then update my local machine, Nix will use that version
to generate the URL from which to download the extension. And for some
extensions, the zip files from /latest are different than the ones from
/$version.