Actually use the nightly overlay
I've had the Neovim nightly overlay included in my config for years and
never noticed that I wasn't using it. Or maybe I was using it and then
broke it somehow.

I'm still not entirely sure how to best define overlays that are then
used by both things like `pkgs.foo` (where I want to have an overlay for
the `foo` package) but also `programs.neovim` where I want the Neovim
package to come from the overlay.

For now I'm using the explicit `programs.neovim.package` option, but I
think that that's not strictly necessary.

https://discourse.nixos.org/t/home-managers-nixpkgs-overlays-doesnt-have-effect/49528

I didn't bother to really research all of this. As long as it works (for
now).