Switch to Telescope
I've been struggling with getting the colors to look right in FzfLua
because it essentially runs a shell program in a floating window.
Telescope seems to work good enough out of the box. It also has the
community power behind it, so I'll hope it's a good choice in the long
run.