Remove NixOS VM again
Working in a VM is super clunky. Clipboard integration is wonky, it
often crashes when my MBP goes to sleep, resolution is always an issue
when switching monitors.

If I need to debug something in a Linux environment I should really try
to do it through Docker or just find some other way (production
systems?)